In this version of the code, we use the ... syntax in the function signature to define a rest parameter named args. This parameter collects all of the arguments passed to the function into an array.
Inside the function body, we simply return the length of the args array, which represents the number of arguments that were passed to the function.

Now, when you call the returnHowManyArguments function with any number of arguments, it will return the number of arguments that were passed. For example, calling returnHowManyArguments("Hello", "Holberton", 2020); will return 3.