Heres how you can modify the concatArrays function to concatenate 2 arrays and each character of a string using spread syntax:
In this version of the code, we use the ... spread syntax to spread the elements of array1, array2, and string into a new array. This creates a new array that contains all of the elements from array1 and array2, followed by each character from string.

For example, calling concatArrays([1, 2], [3, 4], "567") will return [1, 2, 3, 4, "5", "6", "7"].

Received message. Heres how you can modify the concatArrays function to concatenate 2 arrays and each character of a string using spread syntax: javascript export default function concatArrays(array1, array2, string) { return [...array1, ...array2, ...string]; }  In this version of the code, we use the ... spread syntax to spread the elements of array1, array2, and string into a new array. This creates a new array that contains all of the elements from array1 and array2, followed by each character from `string. For example, calling concatArrays([1, 2], [3, 4], 567)` will return `[1, 2, 3, 4, "5", "6", "7"].