In this version of the code, we use computed property names to define the properties of the budget object. Computed property names allow us to use an expression (enclosed in square brackets) to compute the name of a property.
In this case, we use template literals to create property names that include the current year (as returned by the getCurrentYear function). This allows us to create dynamic property names based on the current year.

The resulting budget object will have properties with names like income-2023, gdp-2023, and capita-2023, where 2023 is the current year. The values of these properties will be set to the values of the income, gdp, and capita parameters, respectively.