In this version of the code, we use a for...of loop to iterate over the values in the input array. This allows us to access each value directly, without having to use an index variable like idx.
We also replace the var keyword with const to declare our variables. const is an ES6 keyword that creates block-scoped variables that cannot be reassigned. This makes our code more predictable and easier to understand.

Inside the loop, we use the push method to add each modified value to a new result array. This way, we avoid modifying the input array and instead return a new array with the modified values.

For example, calling appendToEachArrayValue(["a", "b", "c"], "1") will return a new array ["1a", "1b", "1c"], while the input array ["a", "b", "c"] remains unchanged.