In this example, the Teacher interface is defined with the required attributes firstName, lastName, fullTimeEmployee, and location, as well as an optional yearsOfExperience attribute and an index signature that allows for any additional attribute to be added to the object.
A TeacherClass class is then defined that implements the Teacher interface. The class constructor takes in the required attributes as arguments, as well as an optional additionalAttributes parameter that allows for any additional attribute to be passed in as an object.

Inside the constructor, the required attributes are assigned to the class instance using the this keyword. If additionalAttributes are passed in, each attribute is assigned to the instance using a forEach loop and the index signature.

Finally, two instances of the TeacherClass are created with different sets of attributes, demonstrating how the class can be used to create objects with varying sets of attributes.