Teacher Interface: Here is an example of a Teacher interface in a namespace named Subjects that requires firstName and lastName as strings:
Subject.ts Interface:
Here is an example of a Subject class in the same namespace named Subjects that has one attribute teacher that implements the Teacher interface and one setter method setTeacher that accepts a teacher in argument:

modifications in the same namespace named Subjects:
In this example, we use declaration merging to add a new optional attribute experienceTeachingC (number) to the Teacher interface. We also create a class Cpp that extends from Subject. The Cpp class has two methods: getRequirements and getAvailableTeacher. The getRequirements method returns a string "Here is the list of requirements for Cpp". The getAvailableTeacher method returns a string "Available Teacher: <first name of teacher>" if the teacher has experience teaching C. If the teacher doesnt have any experience in teaching C, then the method returns a string "No available teacher".

React.ts Class:
In this example, we use declaration merging to add a new optional attribute experienceTeachingReact (number) to the Teacher interface. We also create a class React that extends from Subject. The React class has two methods: getRequirements and getAvailableTeacher. The getRequirements method returns a string "Here is the list of requirements for React". The getAvailableTeacher method returns a string "Available Teacher: <first name of teacher>" if the teacher has experience teaching React. If the teacher doesnt have any experience in teaching React, then the method returns a string "No available teacher".

Java.ts class:
In this example, we use declaration merging to add a new optional attribute experienceTeachingJava (number) to the Teacher interface. We also create a class Java that extends from Subject. The Java class has two methods: getRequirements and getAvailableTeacher. The getRequirements method returns a string "Here is the list of requirements for Java". The getAvailableTeacher method returns a string "Available Teacher: <first name of teacher>" if the teacher has experience teaching Java. If the teacher doesnt have any experience in teaching Java, then the method returns a string "No available teacher".